+++ apps/crews/rollup.config.cjs
@@ -2,7 +2,6 @@ const nodeResolve = require("@rollup/plugin-node-resolve");
 const commonjs = require("@rollup/plugin-commonjs");
 const typescript = require("rollup-plugin-typescript2");
 const json = require("@rollup/plugin-json");
-const nodePolyfills = require('rollup-plugin-node-polyfills');
 
 /**
  * @type {import("rollup").RollupOptions[]}
@@ -11,7 +10,7 @@ module.exports = [
   {
     input: "src/index.ts",
     output: {
-      dir: "dist/bundle-node",
+      dir: "dist/bundle",
       format: "commonjs",
     },
     plugins: [
@@ -20,22 +19,5 @@ module.exports = [
       commonjs(),
       json(),
     ],
-  },
-  {
-    input: "src/index.ts",
-    output: {
-      dir: "dist/bundle-browser",
-      format: "commonjs",
-    },
-    plugins: [
-      nodeResolve({
-        browser: true,
-        modulesOnly: true,
-      }),
-      typescript(),
-      commonjs(),
-      json(),
-      nodePolyfills()
-    ],
-  },
+  }
 ];