+++ apps/montelo-core/rollup.config.cjs
@@ -2,7 +2,7 @@ const nodeResolve = require("@rollup/plugin-node-resolve");
 const commonjs = require("@rollup/plugin-commonjs");
 const typescript = require("rollup-plugin-typescript2");
 const json = require("@rollup/plugin-json");
-const nodePolyfills = require('rollup-plugin-node-polyfills');
+const nodePlyfills = require('rollup-plugin-node-polyfills');
 
 /**
  * @type {import("rollup").RollupOptions[]}
@@ -35,7 +35,6 @@ module.exports = [
       typescript(),
       commonjs(),
       json(),
-      nodePolyfills(),
     ],
   },
 ];