+++ apps/montelo-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # montelo
 
+## 1.1.71
+
+### Patch Changes
+
+- y
+
 ## 1.1.70
 
 ### Patch Changes