+++ apps/montelo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # montelo
 
+## 1.1.71
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/core@1.1.71
+  - @montelo/crews@1.1.71
+
 ## 1.1.70
 
 ### Patch Changes